[smartplaylist] move CDatabaseQueryRule to a new file
