+++ project/VS2010Express/XBMC.vcxproj.filters
@@ -2450,6 +2450,9 @@
     <ClCompile Include="..\..\xbmc\dbwrappers\Database.cpp">
       <Filter>dbwrappers</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\xbmc\dbwrappers\DatabaseQuery.cpp">
+      <Filter>dbwrappers</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\xbmc\dbwrappers\dataset.cpp">
       <Filter>dbwrappers</Filter>
     </ClCompile>
@@ -5518,6 +5521,9 @@
       <Filter>video</Filter>
     </ClInclude>
     <ClInclude Include="..\..\xbmc\dbwrappers\Database.h">
+      <Filter>dbwrappers</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\xbmc\dbwrappers\DatabaseQuery.h">
       <Filter>dbwrappers</Filter>
     </ClInclude>
     <ClInclude Include="..\..\xbmc\dbwrappers\dataset.h">
