+++ xbmc/dbwrappers/Makefile
@@ -1,4 +1,5 @@
 SRCS=Database.cpp \
+     DatabaseQuery.cpp \
      dataset.cpp \
      mysqldataset.cpp \
      qry_dat.cpp \
